requirements.txt
Django
djangorestframework
django-cors-headers
djoser
Pillow
stripe
